Reduce the load on the user-agent and server
We are now no longer sending *all* errors to the server when there are
unreasonably many. To avoid a significant denial of service, a queuing system
has been implemented to send errors 'slowly' with a back-off as more and more
errors are sent. The first few errors will be sent very quickly though, so we
do not need to be worried about not seeing important errors when the total
number (per page) is not excessive.